Adds docs and moves functions
Moved all functions related to clustering and classifying to a new file 'segimgs.py'
